+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.3.* (2020-03)
 ------------------
 
+* Add proxy and lazy module.
 * Add find_caller to trace source of config value
 * Refactory ConfigMeta data structure